bridge/requirements-ui.txt
# UI runtime requirements (lightweight)
pywebview>=4.4.1
pythonnet>=3.0.0
pywin32>=306
Pillow>=10.4.0
requests>=2.32.3
urllib3>=2.0.0
PySide6>=6.7.0
psutil>=7.1.3
